re-constructing downloader. next work: delete reset

+++ downloader_class.py
@@ -14,54 +14,55 @@ from env import *
 # XXX: change file name for RV when time < Feb, 2003.
 TEST = False
 
+# FIXME not flexible
+def combine_flist(self, order):
+    sdate = daterange[order][0]
+    fnames = {}
+    clist = cmlib.get_collector(sdate)
+    for cl_name in clist:
+        if cl_name.startswith('rrc'):
+            cl_type = 1 # collector type
+        else:
+            cl_type = 0
+        # .bz2.txt.gz file name
+        flist = open(datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name, 'r')
+        for filename in flist:
+            filename = filename.replace('\n', '')
+            file_attr = filename.split('.')
+            if cl_type == 0:
+                file_dt = file_attr[3] + file_attr[4]
+            else:
+                file_dt = file_attr[5] + file_attr[6]
+            dt_obj = datetime.datetime.strptime(file_dt, '%Y%m%d%H%M')
+            fnames[filename] = dt_obj
+        flist.close()
+    fnlist = sorted(fnames, key=fnames.get)
+
+    if os.path.exists(datadir+'metadata/'+sdate+'/updt_filelist_comb'):
+        os.remove(datadir+'metadata/'+sdate+'/updt_filelist_comb')
+    fcomb = open(datadir+'metadata/'+sdate+'/updt_filelist_comb', 'w')  # .bz2.txt.gz file name
+    for fn in fnlist:
+        fcomb.write(fn+'\n')
+    fcomb.close()
+    return 0
+
 class Downloader():
 
-    def __init__(self, order_list):
-        self.order_list = order_list
+    def __init__(self, sdate, edate, co):
+        self.sdate = sdate
+        self.edate = edate
+        self.co = co
 
-    # FIXME not flexible
-    def combine_flist(self, order):
-        sdate = daterange[order][0]
-        fnames = {}
-        clist = cmlib.get_collector(sdate)
-        for cl_name in clist:
-            if cl_name.startswith('rrc'):
-                cl_type = 1 # collector type
-            else:
-                cl_type = 0
-            # .bz2.txt.gz file name
-            flist = open(datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name, 'r')
-            for filename in flist:
-                filename = filename.replace('\n', '')
-                file_attr = filename.split('.')
-                if cl_type == 0:
-                    file_dt = file_attr[3] + file_attr[4]
-                else:
-                    file_dt = file_attr[5] + file_attr[6]
-                dt_obj = datetime.datetime.strptime(file_dt, '%Y%m%d%H%M')
-                fnames[filename] = dt_obj
-            flist.close()
-        fnlist = sorted(fnames, key=fnames.get)
-
-        if os.path.exists(datadir+'metadata/'+sdate+'/updt_filelist_comb'):
-            os.remove(datadir+'metadata/'+sdate+'/updt_filelist_comb')
-        fcomb = open(datadir+'metadata/'+sdate+'/updt_filelist_comb', 'w')  # .bz2.txt.gz file name
-        for fn in fnlist:
-            fcomb.write(fn+'\n')
-        fcomb.close()
-        return 0
-
-    # result: leaving only .bz2/gz.txt.gz update files
-    def parse_update(self, listname): # all update files from one collectors/list
-        flist = open(listname, 'r')
+    # output: .bz2/gz.txt.gz files
+    def parse_updates(self, listfile): # all update files from one collectors/list
+        flist = open(listfile, 'r')
         for line in flist:
-            line = line.split('|')[0]
-            line = line.replace('.txt.gz', '') # get the original .bz2/gz file name
+            line = line.split('|')[0].replace('.txt.gz', '') # get the original .bz2/gz file name
             if not os.path.exists(datadir+line+'.txt.gz'):
                 cmlib.parse_mrt(line, line+'txt') # .bz2/gz => .bz2/gz.txt
                 cmlib.pack_gz(line+'txt') # .bz2/gz.txt => .bz2/gz.txt.gz
                 os.remove(line)  # remove the original .bz2/.gz file
-            else:  # bz2/gz.txt.gz file already exists
+            else:
                 pass
         flist.close()
         return 0
@@ -160,53 +161,59 @@ class Downloader():
 
         return peers
 
-    def get_update_list(self, order, clctr):
-        try:
-            if int(daterange[order][0]) < int(clctr[2]):
-                print 'error: this collector started too late'
-                return -1
-        except:  # usually when testing, clctr[2] may not be set
-            pass
+    def get_update_list(self, clctr, listfile):
+        if int(self.sdate) < int(all_collectors[clctr]):
+            print 'error: this collector started too late'
+            logging.error('collector\' start date too late')
+            return -1
 
-        cl_name = clctr[0]
-        cl_type = clctr[1]
-
-        if cl_type == 0: # RouteViews
-            datadir_detail = datadir + 'archive.routeviews.org/' + cl_name +\
-                '/bgpdata/'
-            datadir_detail = datadir_detail.replace('//', '/') # happens when cl = ''
+        # TODO
+        ''' Seems useless anymore
+        if clctr.startswith('rrc'):
+            datadir_detail = datadir + 'data.ris.ripe.net/' + clctr + '/'
         else:
-            datadir_detail = datadir + 'data.ris.ripe.net/' + cl_name + '/'
+            datadir_detail = datadir + 'archive.routeviews.org/' + clctr + '/bgpdata/'
+            datadir_detail = datadir_detail.replace('//', '/') # when cl = ''
+        '''
 
-        sdate = daterange[order][0]
-        sdate_obj = datetime.datetime.strptime(sdate, '%Y%m%d').date()
-        edate_obj = sdate_obj + datetime.timedelta(days=(daterange[order][1]-1))
-        edate = edate_obj.strftime('%Y%m%d')
+        # change into month quantity for easy calculation
+        smonth = int(sdate[0:4])*12 + int(sdate[4:6])
+        emonth = int(edate[0:4])*12 + int(edate[4:6])
+        month_gap = emonth - smonth # could be zero
 
-        # Now we can only deal with at most 2 months
-        # TODO handle more months
-        yearmonth = [] 
-        yearmonth.append(sdate[0:4] + '.' + sdate[4:6])
-        if edate[0:4] + '.' + edate[4:6] not in yearmonth:
-            yearmonth.append(edate[0:4] + '.' + edate[4:6])
-
-        cmlib.make_dir(datadir+'metadata/'+sdate)
-        if os.path.exists(datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name):
-            os.remove(datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name)
-        flist = open(datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name, 'w')  
-
-        for ym in yearmonth:
-            web_location = ''
-            if cl_type == 0:
-                #web_location = 'archive.routeviews.org/' +\
-                web_location = 'routeviews.org/' +\
-                        cl_name + '/bgpdata/' + ym + '/UPDATES/'
-                web_location = web_location.replace('//', '/')  # when name is ''
-                webraw = cmlib.get_weblist('http://' + web_location)
+        month_list = [] 
+        for m in xrange(0, month_gap+1):
+            now_month = smonth + m
+            if now_month % 12 == 0: # the 12th month
+                now_year = now_month / 12 - 1
+                now_month = 12
             else:
-                web_location = 'data.ris.ripe.net/'+cl_name+'/'+ym+'/' 
-                webraw = cmlib.get_weblist('http://' + web_location)
+                now_year = now_month / 12
+                now_month = now_month % 12
 
+            if now_month / 10 == 0:
+                str_now_month = '0' + str(now_month) # E.g., '5'=>'05'
+            else:
+                str_now_month = str(now_month)
+
+            month_list.append(str(now_year)+'.'+str_now_month) 
+
+
+        tmp_filename = listfile.split('/')[-1]
+        tmp_dir = listfile.replace(tmp_filename, '')
+        cmlib.make_dir(tmp_dir)
+        flist = open(listfile, 'w')  
+
+        for month in month_list:
+            web_location = ''
+            if clctr.startswith('rrc'):
+                web_location = 'data.ris.ripe.net/'+clctr+'/'+month+'/' 
+            else:
+                #web_location = 'archive.routeviews.org/' +\ # XXX I don't know why change
+                web_location = 'routeviews.org/' + clctr + '/bgpdata/' + month + '/UPDATES/'
+                web_location = web_location.replace('//', '/')  # when name is ''
+
+            webraw = cmlib.get_weblist('http://' + web_location)
             cmlib.make_dir(datadir+web_location)
 
             for line in webraw.split('\n'):
@@ -216,40 +223,36 @@ class Downloader():
                 size = line.split()[-1]
                 if size.isdigit():
                     fsize = float(size)
-                else:
+                else: # E.g., 155 K
                     fsize = float(size[:-1]) * cmlib.size_u2v(size[-1])
                 filename = line.split()[0]  # omit uninteresting info
                 filedate = filename.split('.')[-3]
 
-                # check whether its datetime in our range
-                if int(filedate) < int(sdate) or int(filedate) > int(edate):
+                # check whether its date in our range
+                if int(filedate) < int(self.sdate) or int(filedate) > int(self.edate):
                     continue
                 # .bz2/gz.txt.gz file list
                 flist.write(web_location+filename+'.txt.gz|'+str(fsize)+'\n')
-                print web_location+filename+'.txt.gz|'+str(fsize)
+                logging.info('record file name: '+web_location+filename+'.txt.gz|'+str(fsize))
 
-        return datadir+'metadata/'+sdate+'/updt_filelist_'+cl_name # listname
+        return 0
 
 
-    def get_update(self, listname):
-        f = open(listname, 'r')
-        if TEST: # Just read several files when testing
+    def download_updates(self, listfile):
+        if TEST: # XXX Just read several files when testing
             testcount = 0
+
+        f = open(listfile, 'r')
         for line in f:
-            line = line.replace('\n', '')
-            line = line.replace('.txt.gz', '') # for getting original .bz2/gz name
-            tmp = line.split('|')[0] # == web location + filename
+            line = line.replace('\n', '').replace('.txt.gz', '') # get original .bz2/gz name
             filename = tmp.split('/')[-1]
             web_location = tmp.replace(filename, '') 
             fsize = int(line.split('|')[1])
 
-            # remove existing xx.txt file to make things clearer
-            try:
+            # remove (if) existing xx.txt file to make things clearer
+            # consequence: only XXX.bz2/.gz or XXX.bz2/gz.txt.gz exists
+            if os.path.exists(web_location+filename+'.txt'):
                 os.remove(web_location+filename+'.txt')
-            except:
-                pass
-
-            # consequence: ???.bz2/.gz or ???.bz2/gz.txt.gz
 
             if os.path.exists(web_location+filename+'.txt.gz'): # parsed file exists
                 if os.path.getsize(web_location+filename+'.txt.gz') > 0.1 * fsize: # size OK
@@ -269,57 +272,52 @@ class Downloader():
             print 'Downloading ' + 'http://'+web_location + filename
             logging.info('Downloading ' + 'http://'+web_location + filename)
 
-            if TEST:
+            if TEST: # XXX only download 5 files when testing
                 testcount += 1
                 if testcount == 5:
                     break
-
         f.close()
 
         return 0
 
-    ## update and RIB file formats are either .bz2/gz or .xx.txt.gz!
-    def get_file(self):
-        for order in self.order_list:
-            for clctr in collectors:
-                listname = self.get_update_list(order, clctr)
-                if listname == -1: # fail
-                    continue
-                #self.get_update(listname)
-                #self.parse_update(listname)
-                #rib_comp_loc = self.get_parse_rib(order, clctr)
-                #self.delete_reset(rib_comp_loc, listname) # should after RIB download and parse
-            self.combine_flist(order)
+    # Note: update and RIB file formats are either .bz2/gz or .xx.txt.gz!
+    def get_all_updates(self, listfile):
+        for clctr in self.co:
+            listname = self.get_update_list(clctr, listfile)
+            if listname == -1: # fail to create
+                continue
+            self.download_updates(listfile)
+            self.parse_updates(listfile)
+            rib_full_loc = self.get_parse_one_rib(clctr)
+            #self.delete_reset(rib_full_loc, listname) # should after RIB download and parse
+        #self.combine_flist(order)
 
         return 0
 
-    def get_parse_rib(self, order, clctr):
-        cl_name = clctr[0]
-        cl_type = clctr[1]
-        if cl_type == 0:
-            #filelocation = 'archive.routeviews.org/' +\  # RV abandoned this
-            filelocation = 'routeviews.org/' +\
-                    cl_name + '/bgpdata/' + yearmonth[0] + '/RIBS/'
-            filelocation = filelocation.replace('//', '/')  # when name is ''
-            webraw = cmlib.get_weblist('http://' + filelocation)
+    def get_parse_one_rib(self, clctr):
+        tmp_month = self.sdate[0:4] + self.sdate[4:6]
+        if clctr.startswith('rrc'):
+            filelocation = 'data.ris.ripe.net/' + cl_name + '/' + tmp_month + '/' 
         else:
-            filelocation = 'data.ris.ripe.net/' + cl_name + '/' + yearmonth[0] + '/' 
-            webraw = cmlib.get_weblist('http://' + filelocation)
-        
-        cmlib.make_dir(datadir+filelocation)
-        sdate = daterange[order][0]
+            filelocation = 'routeviews.org/' + cl_name + '/bgpdata/' + tmp_month + '/RIBS/'
+            filelocation = filelocation.replace('//', '/')
+        webraw = cmlib.get_weblist('http://' + filelocation)
 
-        # for each event, we only download one RIB (on or near the sdate)
+        cmlib.make_dir(datadir+filelocation)
+
+        # download one RIB on or near sdate only for deleting reset
+        # TODO: if duration too long, we may need multiple RIBs
         rib_list = webraw.split('\n')
         filter(lambda a: a != '', rib_list)
         filter(lambda a: a != '\n', rib_list)
         rib_list = [item for item in rib_list if 'rib' in item or 'bview' in item]
 
-        target_line = ''
+        # TODO: avoid the RIB with strange size
+        target_line = '' # the RIB file for downloading
         closest = 99999
         for line in rib_list:
             fdate = line.split()[0].split('.')[-3]
-            diff = abs(int(fdate)-int(sdate)) # >0
+            diff = abs(int(fdate)-int(self.sdate)) # >0
             if diff < closest:
                 closest = diff
                 target_line = line
@@ -331,34 +329,32 @@ class Downloader():
             fsize = float(size[:-1]) * cmlib.size_u2v(size[-1])
 
         filename = target_line.split()[0]
-        complete_loc = datadir + web_location + filename # .bz2/.gz
+        full_loc = datadir + web_location + filename # .bz2/.gz
 
-        try:
-            os.remove(complete_loc+'.txt')
-        except:
-            pass
+        if os.path.exists(full_loc+'.txt'):
+            os.remove(full_loc+'.txt')
 
-        if os.path.exists(complete_loc+'.txt.gz'): 
-            if os.path.getsize(complete_loc+'.txt.gz') > 0.1 * fsize:
-                if os.path.exists(complete_loc):  # .bz2/.gz useless anymore
-                    os.remove(complete_loc)
-                return complete_loc+'.txt.gz'
+        if os.path.exists(full_loc+'.txt.gz'): 
+            if os.path.getsize(full_loc+'.txt.gz') > 0.1 * fsize:
+                if os.path.exists(full_loc):  # .bz2/.gz useless anymore
+                    os.remove(full_loc)
+                return full_loc+'.txt.gz'
             else:
-                os.remove(complete_loc+'.txt.gz') # too small to be complete
+                os.remove(full_loc+'.txt.gz') # too small to be complete
                 cmlib.force_download_file('http://'+web_location, datadir+web_location, filename)
 
-        if os.path.exists(complete_loc): 
-            if os.path.getsize(complete_loc) <= 0.9 * fsize:
-                os.remove(complete_loc)
+        if os.path.exists(full_loc): 
+            if os.path.getsize(full_loc) <= 0.95 * fsize:
+                os.remove(full_loc)
                 cmlib.force_download_file('http://'+web_location, datadir+web_location, filename)
             else:
                 pass
 
-        cmlib.parse_mrt(complete_loc, complete_loc+'.txt')
-        os.remove(complete_loc)  # then remove .bz2/.gz
-        cmlib.pack_gz(complete_loc+'.txt')
+        cmlib.parse_mrt(full_loc, full_loc+'.txt')
+        os.remove(full_loc)  # then remove .bz2/.gz
+        cmlib.pack_gz(full_loc+'.txt')
 
-        return complete_loc+'.txt.gz'
+        return full_loc+'.txt.gz'
 
 
     def delete_reset(self, rib_comp_loc, listname):
@@ -392,16 +388,14 @@ class Downloader():
         return 0
 
 if __name__ == '__main__':
-    '''
-    order_list = [0, 1]
-    collector_list = []
+    order_list = [27]
+    collector_list = ['', 'rrc00']
     for order in order_list:
-        sdate = 
-        edate = 
+        sdate = daterange[order][0]
+        edate = daterange[order][1]
         for co in collector_list:
             dl = Downloader(sdate, edate, co)
             listfile = datadir + 'update_list/' + sdate + '_' + edate + '/' + co + '_list.txt'
             dl.get_all_updates(listfile) # para: update list file location
-    '''
-    dl = Downloader([0])
-    dl.get_file()
+    #dl = Downloader([0])
+    #dl.get_file()
